drm/vs: Add Base API for VS Mode Configuration
This commit adds a base API for configuring VS modes,
which will streamline the setup and management of display modes
in the VS DRM subsystem.

In this implementation, we are using drm_atomic_helper_commit_tail_rpm()
instead of drm_atomic_helper_commit_tail() to ensure that
we skip planes related to inactive CRTCs.

This helps to optimize the commit process and reduces unnecessary overhead
when dealing with inactive display resources.

Signed-off-by: keith zhao <keith.zhao@starfivetech.com>